Tukat int kepada size_t

+++ string.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int my_strlen(char *s)
+size_t my_strlen(char *s)
 {
     char *p = s;
     while (*p)
